requirements.txt
gspread
requests
python-dotenv
shipstation